+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Added config to control whether other players should receive join/leave messages when a player (un-)vanishes
 - Added configs to change the messages the player receives when vanishing/unvanishing
+- Added subtitle message displaying the current vanished status when vanishing/unvanishing or querying the vanished status
 - Fixed that Command Blocks are not able to execute /v by default
 
 -----------Changelog for v1.1.2-----------